поправил синтаксис

+++ BookWorm/Construct/ConstructCellBorder.cs
@@ -51,7 +51,6 @@ namespace BookWorm.Construct
         /// <param name="DA">The DA object is used to retrieve from inputs and store in outputs.</param>
         protected override void SolveInstance(IGH_DataAccess DA)
         {
-            
             var style = 0;
             var color = string.Empty;
             var border = new Border();
@@ -65,14 +64,12 @@ namespace BookWorm.Construct
             {
                 border.Color = new Color
                 {
-                    Alpha = (float)255,
+                    Alpha = 255,
                     Red = float.Parse(color.Split(',')[0]),
                     Green = float.Parse(color.Split(',')[1]),
                     Blue = float.Parse(color.Split(',')[2]),
                 };
             }
-
-            
 
             var borderGoo = new GH_CellBorder(border);
             DA.SetData(0, borderGoo);
@@ -81,6 +78,7 @@ namespace BookWorm.Construct
         /// <summary>
         /// Gets provides an Icon for the component.
         /// </summary>
+
         protected override System.Drawing.Bitmap Icon
         {
             get
